+++ Proyecto_Final/src/entidad/Cliente.java
@@ -2,7 +2,7 @@ package entidad;
 
 public class Cliente {
     private String nombre;
-    private int telefono;
+    private long telefono;
 
     public String getNombre() {
         return nombre;
@@ -12,18 +12,18 @@ public class Cliente {
         this.nombre = nombre;
     }
 
-    public int getTelefono() {
+    public long getTelefono() {
         return telefono;
     }
 
-    public void setTelefono(int telefono) {
+    public void setTelefono(long telefono) {
         this.telefono = telefono;
     }
 
     public Cliente() {
     }
 
-    public Cliente(String nombre, int telefono) {
+    public Cliente(String nombre, long telefono) {
         this.nombre = nombre;
         this.telefono = telefono;
     }
