+++ Proyecto_Final/src/entidad/Venta.java
@@ -1,14 +1,14 @@
 package entidad;
 
 public class Venta {
-    private float precio;
+    private double precio;
     private String formaPago;
 
-    public float getPrecio() {
+    public double getPrecio() {
         return precio;
     }
 
-    public void setPrecio(float precio) {
+    public void setPrecio(double precio) {
         this.precio = precio;
     }
 
@@ -23,7 +23,7 @@ public class Venta {
     public Venta() {
     }
 
-    public Venta(float precio, String formaPago) {
+    public Venta(double precio, String formaPago) {
         this.precio = precio;
         this.formaPago = formaPago;
     }
